requirements.txt
annotated-types==0.7.0
anyio==4.10.0
Authlib==1.6.1
aws-xray-sdk==2.14.0
aws_lambda_powertools==3.19.0
bandit==1.8.6
black==25.1.0
boto3==1.40.11
botocore==1.40.11
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.2.1
coverage==7.10.3
cryptography==45.0.6
distlib==0.4.0
dparse==0.6.4
filelock==3.16.1
flake8==7.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.13
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy==1.17.1
mypy_extensions==1.1.0
nltk==3.9.1
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
pbr==7.0.0
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.3.0
psutil==6.1.1
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.1
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
s3transfer==0.13.1
safety==3.6.0
safety-schemas==0.0.14
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stevedore==5.4.1
tabulate==0.9.0
tenacity==9.1.2
tomlkit==0.13.3
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
virtualenv==20.34.0
wrapt==1.17.3
